feat: account details
