+++ app/src/main/java/com/dscvit/vitty/activity/AuthActivity.kt
@@ -12,8 +12,6 @@ import androidx.viewpager2.widget.ViewPager2
 import com.dscvit.vitty.R
 import com.dscvit.vitty.adapter.IntroAdapter
 import com.dscvit.vitty.databinding.ActivityAuthBinding
-import com.dscvit.vitty.util.Constants.EMAIL
-import com.dscvit.vitty.util.Constants.NAME
 import com.dscvit.vitty.util.Constants.TOKEN
 import com.dscvit.vitty.util.Constants.UID
 import com.dscvit.vitty.util.Constants.USER_INFO
@@ -127,10 +125,9 @@ class AuthActivity : AppCompatActivity() {
         loginClick = false
     }
 
-    private fun saveInfo(name: String?, email: String?, token: String?, uid: String?) {
+    private fun saveInfo(token: String?, uid: String?) {
         with(sharedPref.edit()) {
-            putString(NAME, name)
-            putString(EMAIL, email)
+            putString("sign_in_method", "Google")
             putString(TOKEN, token)
             putString(UID, uid)
             apply()
@@ -158,7 +155,7 @@ class AuthActivity : AppCompatActivity() {
             if (it.isSuccessful) {
                 loginClick = true
                 val uid = firebaseAuth.currentUser?.uid
-                saveInfo(acct.displayName, acct.email, acct.idToken, uid)
+                saveInfo(acct.idToken, uid)
                 val intent = Intent(this, InstructionsActivity::class.java)
                 binding.loadingView.visibility = View.GONE
                 startActivity(intent)
