+++ app/src/main/java/com/dscvit/vitty/model/UserDetails.kt
@@ -1,3 +1,11 @@
 package com.dscvit.vitty.model
 
-data class UserDetails(var name: String? = "", var email: String? = "", var token: String? = "")
+import android.net.Uri
+
+data class UserDetails(
+    var name: String? = "",
+    var email: String? = "",
+    var photoUrl: Uri? = null,
+    var token: String? = "",
+    var method: String? = "Google",
+)
