+++ app/src/main/res/values/strings.xml
@@ -25,16 +25,16 @@
     <string name="today_schedule">Today\'s Schedule</string>
     <string name="share">Share</string>
     <string name="share_text">Download the VITTY extension and app now!\nExtension: https://dscv.it/vitty-extension\nAndroid App: https://dscv.it/app/vitty</string>
-    <string name="instructions_1">1. Download the Chrome Extension from:</string>
-    <string name="instructions_1_link"><u>https://dscv.it/vitty-extension</u></string>
-    <string name="instructions_2">2. Log in with the same Google Account as shown above.</string>
-    <string name="instructions_3">3. Upload a screenshot of your timetable.</string>
+    <string name="instructions_1">1. Upload the timetable on:</string>
+    <string name="instructions_1_link"><u>https://dscv.it/vitty</u></string>
+    <string name="instructions_2">2. Log in with the same account as shown above.</string>
+    <string name="instructions_3">3. Follow the instructions and upload your timetable.</string>
     <string name="instructions_3_a">\u25CB Ensure that the screenshot is taken directly from VTOP. It should fit in one single image.</string>
     <string name="instructions_3_b">\u25CB DO NOT upload a screenshot that has been compressed or sent through social media.</string>
     <string name="instructions_4">4. Review it.</string>
     <string name="instructions_4_a">\u25CB Individual slot data can be edited by clicking on the edit button.</string>
     <string name="instructions_4_b">\u25CB If unsatisfied, click on the three dots and select Upload as Text.</string>
-    <string name="instructions_5">5. When done, click on Upload.</string>
+    <string name="instructions_5">5. When done, click on Confirm.</string>
     <string name="instructions_6">BRAVO! That\'s it. You did it!</string>
     <string name="instructions_7">In addition to this, you can have a sneak peek at your upcoming classes via the Widget.</string>
     <string name="logout">Logout</string>
@@ -47,7 +47,7 @@
     <string name="battery">Battery</string>
     <string name="nav">Navigation</string>
     <string name="new_updates">What\'s New?</string>
-    <string name="account_details">Account Details</string>
+    <string name="account_details">Google Account Details</string>
     <string name="notif_group">Courses</string>
     <string name="complete_msg">You have successfully set up your timetable! You deserve a pat on your back!</string>
     <string name="skip">Skip</string>
